Auto detected dev specs will be minimal required (#500)
* Auto detected dev specs will be minimal required

Currently spack-manager just copies the entire spec you give it when
using create-dev-env.  Now it will strip it to the minimal required spec
of `{name}@{version}`. This is to make it easier to read, edit, and
encourage matches.  Also a recent change in behavior in spack was detected
by @alanw0 in which full specs to fail some times. We do not yet
understand what caused that to occur.

* A minor fix in the docs for readability