Distribution extra spack cleanup (#640)
* Minor fixes

fix tests to be robust to non-breaking api changes.
Also, change binary asset check so it is not recursive.
This avoids an infinite loop in some environments.

* When bundling spack with a distribution don't include user files

The distribution command was including files specific to the user's spack environment
were getting bundled along with spack.  This is becasue spack writes into its root
directory as it is being used.  This creates a more exhaustive list of locations
to avoid copying along with spack and more agressively asserts they don't end up in
the final spack installation